+++ src/components/Timer/PrayerTime.js
@@ -1,12 +1,11 @@
 import React, {useState, useEffect} from 'react';
-import angel from '../../assets/images/Prayer-time.jpg';
 import LocationText from './LocationText';
-import {useRenderCounts} from '../../hooks/api-hooks';
+import {useRenderCounts, useCmsAsset} from '../../hooks/api-hooks';
 import Bgmusic from '../Prayers/Bgmusic'
-import bgm from '../../assets/mp3/quietTime.mp3'
+import bgm from '../../assets/mp3/quietTime.mp3';
+import {PRISMIC_PRAYERTIME_BOY_BG} from '../../utils/constants';
 
 const styles = {
-    backgroundImage:`url(${angel})`, 
     height:'250px', 
     backgroundSize: 'auto 100%', 
     backgroundRepeat:'no-repeat', 
@@ -22,6 +21,7 @@ const styles = {
 
 
 const PrayerTime = ({anim, setAnim, travel=false, location=''}) => {
+    const angelimage = useCmsAsset(PRISMIC_PRAYERTIME_BOY_BG)
     useRenderCounts('PrayerTime.js');
     const [music, setMusic] = useState({show: false, playing : false});
     const [volume, setVolume] = React.useState(true);
@@ -38,7 +38,7 @@ const PrayerTime = ({anim, setAnim, travel=false, location=''}) => {
         <Bgmusic bgm={bgm} volume={volume} setPlaying={() => {setMusic({show: true, playing : true})}}/>
         <div style={{overflow: 'hidden'}}>                
                 {anim[0] && <div className="Prayer-time" 
-                    style={styles}>
+                    style={{...styles, backgroundImage:`url(${angelimage})`}}>
                         {anim[1]} Time
                         {travel && <>
                             <br />